refactor: move isVailedJSON import to utils for better organization
