+++ src/editMessageString.tsx
@@ -1,5 +1,5 @@
 import { ChatStore, ChatStoreMessage } from "@/types/chatstore";
-import { isVailedJSON } from "@/message";
+import { isVailedJSON } from "@/utils/isVailedJSON";
 import { calculate_token_length } from "@/chatgpt";
 import { Tr } from "@/translate";
